Consistency in handling error when fetching projection status
