+++ src/Projection/PdoEventStoreReadModelProjector.php
@@ -24,6 +24,7 @@ use Prooph\Common\Messaging\Message;
 use Prooph\EventStore\EventStore;
 use Prooph\EventStore\EventStoreDecorator;
 use Prooph\EventStore\Exception;
+use Prooph\EventStore\Pdo\Exception\RuntimeException;
 use Prooph\EventStore\Pdo\MySqlEventStore;
 use Prooph\EventStore\Pdo\PostgresEventStore;
 use Prooph\EventStore\Projection\ProjectionStatus;
@@ -465,7 +466,16 @@ EOT;
         try {
             $statement->execute([$this->name]);
         } catch (PDOException $exception) {
-            // ignore
+            // ignore and check error code
+        }
+
+        if ($statement->errorCode() !== '00000') {
+            $errorCode = $statement->errorCode();
+            $errorInfo = $statement->errorInfo()[2];
+
+            throw new RuntimeException(
+                "Error $errorCode. Maybe the projection table is not setup?\nError-Info: $errorInfo"
+            );
         }
 
         $result = $statement->fetch(PDO::FETCH_OBJ);
