+++ tests/Projection/PdoEventStoreReadModelProjectorTest.php
@@ -191,30 +191,4 @@ abstract class PdoEventStoreReadModelProjectorTest extends AbstractEventStoreRea
             10
         );
     }
-
-    /**
-     * @test
-     */
-    public function it_handles_missing_projection_table(): void
-    {
-        $this->expectException(\Prooph\EventStore\Exception\RuntimeException::class);
-        $this->expectExceptionMessage('Maybe the projection table is not setup?');
-
-        $this->prepareEventStream('user-123');
-
-        $this->connection->exec('DROP TABLE projections;');
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', new ReadModelMock());
-
-        $projection
-            ->fromStream('user-123')
-            ->when([
-                UserCreated::class => function (array $state, UserCreated $event): array {
-                    $this->stop();
-
-                    return $state;
-                },
-            ])
-            ->run();
-    }
 }
