+++ tests/Projection/PostgresEventStoreReadModelProjectorTest.php
@@ -16,6 +16,8 @@ use Prooph\Common\Messaging\FQCNMessageFactory;
 use Prooph\EventStore\Pdo\PersistenceStrategy\PostgresSimpleStreamStrategy;
 use Prooph\EventStore\Pdo\PostgresEventStore;
 use Prooph\EventStore\Pdo\Projection\PostgresProjectionManager;
+use ProophTest\EventStore\Mock\ReadModelMock;
+use ProophTest\EventStore\Mock\UserCreated;
 use ProophTest\EventStore\Pdo\TestUtil;
 
 /**
@@ -42,5 +44,31 @@ class PostgresEventStoreReadModelProjectorTest extends PdoEventStoreReadModelPro
             $this->eventStore,
             $this->connection
         );
+    }
+
+    /**
+     * @test
+     */
+    public function it_handles_missing_projection_table(): void
+    {
+        $this->expectException(\Prooph\EventStore\Pdo\Exception\RuntimeException::class);
+        $this->expectExceptionMessage("Error 42P01. Maybe the projection table is not setup?\nError-Info: ERROR:  relation \"projections\" does not exist\nLINE 1: SELECT status FROM projections WHERE name = $1 LIMIT 1;");
+
+        $this->prepareEventStream('user-123');
+
+        $this->connection->exec('DROP TABLE projections;');
+
+        $projection = $this->projectionManager->createReadModelProjection('test_projection', new ReadModelMock());
+
+        $projection
+            ->fromStream('user-123')
+            ->when([
+                UserCreated::class => function (array $state, UserCreated $event): array {
+                    $this->stop();
+
+                    return $state;
+                },
+            ])
+            ->run();
     }
 }
